Alpha v2.1
In this version Gw2DPS will automatically load into "solo mode".
Added a server connect/disconnect feature to the toolbar (red/green
icon)
Reformatted the details section to help clear it up. (still needs more
work)
Added a few more checks to see if you are connected to a server or not.
Renamed letsConnect() to Initialize()